ai-platform/services/frontend/requirements.txt
streamlit>=1.32.0
httpx>=0.27.0
pandas>=2.2.0
pyjwt>=2.8.0
python-dotenv>=1.0.1
